Show the glucose chart in the widget (#377)
Show the glucose chart in the widget. Thanks @bharat!
