+++ Cartfile
@@ -1,6 +1,6 @@
 github "LoopKit/LoopKit" ~> 1.2.0
 github "LoopKit/xDripG5" ~> 0.8.0
-github "i-schuetz/SwiftCharts" ~> 0.5.0
+github "bharat/SwiftCharts" "0.5.1-extension-api-safe"
 github "mddub/dexcom-share-client-swift" ~> 0.2.0
 github "mddub/G4ShareSpy" ~> 0.3.1
 github "ps2/rileylink_ios" ~> 1.0
