+++ Cartfile.resolved
@@ -1,7 +1,7 @@
-github "amplitude/Amplitude-iOS" "v3.14.0"
+github "amplitude/Amplitude-iOS" "v3.14.1"
 github "mddub/G4ShareSpy" "v0.3.1"
 github "LoopKit/LoopKit" "v1.2.0"
-github "i-schuetz/SwiftCharts" "0.5.1"
+github "bharat/SwiftCharts" "0.5.1-extension-api-safe"
 github "mddub/dexcom-share-client-swift" "v0.2.0"
 github "ps2/rileylink_ios" "v1.0"
 github "LoopKit/xDripG5" "v0.8.0"
