+++ Loop/Views/ChartTableViewCell.swift
@@ -7,6 +7,7 @@
 //
 
 import UIKit
+import LoopUI
 
 
 final class ChartTableViewCell: UITableViewCell {
