+++ LoopUI/Views/ChartContentView.swift
@@ -1,0 +1,47 @@
+//
+//  ChartContentView.swift
+//  Loop
+//
+//  Created by Nate Racklyeft on 9/14/16.
+//  Copyright © 2016 Nathan Racklyeft. All rights reserved.
+//
+
+import UIKit
+
+public class ChartContentView: UIView {
+
+    override public func layoutSubviews() {
+        super.layoutSubviews()
+
+        if chartView == nil || chartView!.frame != bounds {
+            chartView = chartGenerator?(bounds)
+        } else if chartView!.superview == nil {
+            addSubview(chartView!)
+        }
+    }
+
+    public func reloadChart() {
+        chartView = nil
+        setNeedsLayout()
+    }
+
+    public var chartGenerator: ((CGRect) -> UIView?)? {
+        didSet {
+            chartView = nil
+            setNeedsLayout()
+        }
+    }
+
+    private var chartView: UIView? {
+        didSet {
+            if let view = oldValue {
+                view.removeFromSuperview()
+            }
+
+            if let view = chartView {
+                self.addSubview(view)
+            }
+        }
+    }
+
+}
